+++ engine/LegacyEngine/Integration/PlayerDevelopmentLevel.cs
@@ -1,0 +1,10 @@
+namespace LegacyEngine.Integration;
+
+public enum PlayerDevelopmentLevel
+{
+    Junior,
+    College,
+    Europe,
+    Ahl,
+    Nhl
+}